requirements.txt
django
channels
pypiwin32
graphene
graphene-django
django-graphql-jwt
django-taggit
django-widget-tweaks
django-filter
django-cors-headers
Pillow
transliterate
gunicorn
psycopg2-binary
graphene-file-upload